+++ dif-data-classes/src/commonMain/kotlin/at/asitplus/dif/FormatHolder.kt
@@ -15,11 +15,6 @@ data class FormatHolder(
     val jwtVp: FormatContainerJwt? = null,
     @SerialName("jwt_vc")
     val jwtVc: FormatContainerJwt? = null,
-    /**
-     * Deprecated, unofficial value, please use [jwtSd]
-     */
-    @SerialName("jwt_sd")
-    val jwtSdDeprecated: FormatContainerJwt? = null,
     @SerialName("vc+sd-jwt")
     val jwtSd: FormatContainerJwt? = null,
     @SerialName("ldp")