+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/OpenIdConstants.kt
@@ -37,6 +37,8 @@ object OpenIdConstants {
 
     const val PATH_WELL_KNOWN_OPENID_CONFIGURATION = "/.well-known/openid-configuration"
 
+    const val PATH_WELL_KNOWN_JWT_VC_ISSUER_METADATA = "/.well-known/jwt-vc-issuer"
+
     const val SCOPE_OPENID = "openid"
 
     const val SCOPE_PROFILE = "profile"