+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/CredentialToBeIssued.kt
@@ -1,30 +1,10 @@
 package at.asitplus.wallet.lib.agent
 
-import at.asitplus.KmmResult
 import at.asitplus.signum.indispensable.CryptoPublicKey
 import at.asitplus.wallet.lib.data.ConstantIndex
 import at.asitplus.wallet.lib.data.CredentialSubject
 import at.asitplus.wallet.lib.iso.IssuerSignedItem
 import kotlinx.datetime.Instant
-
-/**
- * Provides data for credentials to be issued.
- */
-interface IssuerCredentialDataProvider {
-
-    /**
-     * Gets called with a resolved [credentialScheme],
-     * the holder key in [subjectPublicKey] and the requested credential [representation].
-     * Callers may optionally define some attribute names from [ConstantIndex.CredentialScheme.claimNames] in
-     * [claimNames] to request only some claims (if supported by the representation).
-     */
-    fun getCredential(
-        subjectPublicKey: CryptoPublicKey,
-        credentialScheme: ConstantIndex.CredentialScheme,
-        representation: ConstantIndex.CredentialRepresentation,
-        claimNames: Collection<String>? = null,
-    ): KmmResult<CredentialToBeIssued>
-}
 
 sealed class CredentialToBeIssued {
     data class VcJwt(
@@ -49,4 +29,10 @@ sealed class CredentialToBeIssued {
     ) : CredentialToBeIssued()
 }
 
-data class ClaimToBeIssued(val name: String, val value: Any)
+/**
+ * Represents a claim that shall be issued to the holder,
+ * i.e. serialized into the appropriate credential format.
+ * To issue nested structures in SD-JWT, pass a Collection of [ClaimToBeIssued] in [value].
+ * For each claim, one can select if the claim shall be selectively disclosable, or otherwise included plain.
+ */
+data class ClaimToBeIssued(val name: String, val value: Any, val selectivelyDisclosable: Boolean = true)